[3단계 - 위시 리스트] 부산대 BE 김성문 미션 제출합니다. (#322)
* docs: 구현할 기능 목록 정리

* feat: 토큰 전처리를 위한 커스텀 어노테이션과 리졸버 구현

* feat: 위시리스트 상품 목록 조회 기능 구현

* feat: 위시리스트 상품 추가 기능 구현

* feat: 위시리스트 상품 삭제 기능 구현

* fix: 불필요한 로직 제거

* refactor: 데이터베이스 쿼리 로직 최적화

* fix: wishlist 테이블 ddl에 not null 옵션 추가

* test: 서비스 계층 테스트코드 작성

* test: 리포지토리 계층 테스트코드 작성

* test: 컨트롤러 계층 테스트코드 작성

* refactor: 리졸버에 토큰의 멤버id가 유효한지 검사하는 로직 추가

* refactor: wishitem 엔티티 추가 및 데이터베이스 스키마 변경

* fix: bearer prefix 추가

* refactor: rowmapper 구현체 변경

- DataClassRowMapper -> SimplePropertyRowMapper

* refactor: 상품 가격의 타입 변경

- Integer -> Long

* test: 구조 변경에 따른 테스트코드 수정

+++ src/main/java/gift/domain/Product.java
@@ -4,21 +4,21 @@ public class Product {
 
     private final Long id;
     private final String name;
-    private final Integer price;
+    private final Long price;
     private final String imageUrl;
 
-    private Product(Long id, String name, Integer price, String imageUrl) {
+    private Product(Long id, String name, Long price, String imageUrl) {
         this.id = id;
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
     }
 
-    public static Product of(Long id, String name, Integer price, String imageUrl) {
+    public static Product of(Long id, String name, Long price, String imageUrl) {
         return new Product(id, name, price, imageUrl);
     }
 
-    public static Product of(String name, Integer price, String imageUrl) {
+    public static Product of(String name, Long price, String imageUrl) {
         return new Product(null, name, price, imageUrl);
     }
 
@@ -30,7 +30,7 @@ public class Product {
         return name;
     }
 
-    public Integer getPrice() {
+    public Long getPrice() {
         return price;
     }
